+++ lib/pages/admin/promotion/promotion_add.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:salon_app/constants/colors.dart';
-import 'package:salon_app/pages/admin/reservation.dart';
+import 'package:salon_app/pages/admin/reservation/reservation_list.dart';
 import 'package:salon_app/widgets/button.dart';
 
 class AddPromotion extends StatefulWidget {
