+++ lib/pages/admin/services/services_list.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:salon_app/pages/admin/reservation.dart';
+import 'package:salon_app/pages/admin/reservation/reservation_list.dart';
 import 'package:salon_app/widgets/service_section.dart';
-
-import '../../widgets/button2.dart';
+import '../../../widgets/button2.dart';
 
 class ServicesData extends StatelessWidget {
   final String name;
